requirements.txt
numpy 
opencv-python
pillow 
streamlit==1.38.0
streamlit-drawable-canvas==0.9.3